Trabalho 4 pronto

+++ include/TileMap.h
@@ -15,7 +15,7 @@ class TileMap : public Component {
     void Load (std::string file);
     void SetTileSet (TileSet* tileSet);
     int& At (int x, int y, int z = 0);
-    void RenderLayer (int layer, int cameraX = 0, int cameraY = 0);
+    void RenderLayer (int layer, float cameraX = 0, float cameraY = 0);
     int GetWidth ();
     int GetHeight ();
     int GetDepth ();
